don't clobber results

+++ srm_experiments/image_classification_sherlock.py
@@ -4,6 +4,7 @@ import os
 from sklearn.svm import NuSVC, SVC
 import numpy as np
 import models
+from pathlib import Path
 import pickle
 import logging
 import pandas as pd
@@ -37,7 +38,7 @@ def run_experiment(par):
 
     # Run a leave-one-out cross validation with the subjects
     accuracy = np.zeros((subjects,))
-    train_acc  = np.zeros((subjects,)) # upper bound
+    train_acc = np.zeros((subjects,)) # upper bound
 
     for subject in range(subjects):
         
@@ -85,6 +86,8 @@ if __name__ == "__main__":
     # cartesian over param settings
     allpar = [dict(parset) for parset in (zip(runPars.keys(), p)
               for p in product(*runPars.values()))]
+    allpar = [dict(parset) for parset in (zip(runPars.keys(), p)
+              for p in product(*runPars.values()))]
 
     pointsPerId = len(allpar) / totalIDs
     start = int((myID-1)*pointsPerId)
@@ -93,9 +96,12 @@ if __name__ == "__main__":
     # mypar = allpar[start:end]
 
     for parnum in range(start, end):
-        res = pd.DataFrame([run_experiment(allpar[parnum])])
         fname = outfile_template % parnum
-        res.to_csv(fname)
+        if Path(fname).exists(): 
+            continue 
+        else:
+            res = pd.DataFrame([run_experiment(allpar[parnum])])
+            res.to_csv(fname)
 
     print("Done!")
